requirements.txt
aniso8601==8.0.0
attrs==20.3.0
azure-core==1.9.0
azure-storage-blob==12.6.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
flasgger==0.9.5
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==2.10
importlib-metadata==3.1.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
msrest==0.6.19
oauthlib==3.1.0
pycparser==2.20
pyodbc==4.0.30
pyrsistent==0.17.3
pytz==2020.1
PyYAML==5.3.1
repoze.lru==0.7
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0